move bench_one_batch_server logic out of test/ into the entry module
